# server-openapi.yaml
openapi: 3.1.0
info:
  title: FIRMUPS Server API
  version: '0.1.0'
  description: |
    The firmups server API allows for device and update management.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: "http://127.0.0.1:3000"
    description: Local development server
...
paths:
...
  /device_type/{id}:
    get:
      tags:
        - DeviceType
      security:
        - api_key: []
      summary: Get device type
      operationId: getDeviceType
      parameters:
        - name: id
          in: path
          description: ID of the DeviceType to be returned
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceType'
        '404':
          description: Device type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
    patch:
      tags:
        - DeviceType
      security:
        - api_key: []
      summary: Update an existing device type
      operationId: UpdateDeviceType
      parameters:
        - name: id
          in: path
          description: ID of the DeviceType to be updated
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceType'
      responses:
        '200':
          description: Device type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceType'
        '400':
          description: Invalid name provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Device type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Device type already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
    delete:
      tags:
        - DeviceType
      security:
        - api_key: []
      summary: Delete device type
      operationId: deleteDeviceType
      parameters:
        - name: id
          in: path
          description: ID of the DeviceType to be deleted
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceType'
        '404':
          description: Device type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
...
  /device/{id}:
    get:
      tags:
        - Device
      security:
        - api_key: []
      summary: Get device
      operationId: getDevice
      parameters:
        - name: id
          in: path
          description: ID of the Device to be returned
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
    patch:
      tags:
        - Device
      security:
        - api_key: []
      summary: Update an existing device
      operationId: updateDevice
      parameters:
        - name: id
          in: path
          description: ID of the Device to be updated
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDevice'
      responses:
        '200':
          description: Device updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: Invalid input data provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Device already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
    delete:
      tags:
        - Device
      security:
        - api_key: []
      summary: Delete device
      operationId: deleteDevice
      parameters:
        - name: id
          in: path
          description: ID of the Device to be deleted
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
...
components:
  schemas:
...
    UpdateDeviceType:
      type: object
      properties:
        name:
          type: string
    DeviceType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    DeviceStatus:
      type: string
      enum: ["ACTIVE", "INACTIVE", "MAINTENANCE"]
...
    UpdateDevice:
      type: object
      properties:
        name:
          type: string
        type_:
          description: Device type id
          type: integer
        firmware:
          type: integer
        desired_firmware:
          type: integer
        status:
          $ref: '#/components/schemas/DeviceStatus'
    Device:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type_:
          description: Device type id
          type: integer
        firmware:
          type: integer
        desired_firmware:
          type: integer
        status:
          $ref: '#/components/schemas/DeviceStatus'
      required:
        - id
        - name
        - type_
        - firmware
        - desired_firmware
        - status
...
    InternalError:
      description: Masked internal error. The id can be matched with the backend logs.
      type: object
      properties:
        error_id:
          type: string
      required:
        - error_id
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
  securitySchemes:
    api_key:
      type: apiKey
      name: x-api-key
      in: header